+++ module.json
@@ -3,7 +3,7 @@
   "title": "Difficult Terrain",
   "description": "Add a difficult terrain multiplier to your ruler!",
   "author": "Nordiii",
-  "version": "1.0.0",
+  "version": "1.0.1",
   "minimumCoreVersion": "0.6.6",
   "compatibleCoreVersion": "0.6.6",
   "languages": [
@@ -21,7 +21,7 @@
   "esmodules": [
     "scripts/difficult-terrain.js"
   ],
-  "url": "https://github.com/Nordiii/difficultterrain.git",
+  "url": "https://github.com/Nordiii/difficultterrain",
   "manifest": "https://github.com/Nordiii/difficultterrain/releases/latest/download/module.json",
-  "download": "https://github.com/Nordiii/difficultterrain/releases/download/1.0.0/module.zip"
+  "download": "https://github.com/Nordiii/difficultterrain/releases/download/1.0.1/module.zip"
 }
