+++ scripts/difficult-terrain.js
@@ -1,54 +1,6 @@
 import {patch_ruler} from "./patch-ruler.js";
+import {ModuleSettings} from "./module-settings.js";
 
 Hooks.once("ready", () => patch_ruler())
 
-Hooks.once('init', () => {
-    game.settings.register('difficultterrain', "maxTerrainMultiplier", {
-        name: "difficultterrain.maxTerrainMultiplier.n",
-        hint: "difficultterrain.maxTerrainMultiplier.h",
-        scope: "world",
-        config: true,
-        default: 2,
-        type: Number
-    });
-    game.settings.register('difficultterrain', "terrainMultiplierSteps", {
-        name: "difficultterrain.terrainMultiplierSteps.n",
-        hint: "difficultterrain.terrainMultiplierSteps.h",
-        scope: "world",
-        config: true,
-        default: 1,
-        type: Number
-    });
-    game.settings.register('difficultterrain', "incrementSpeed", {
-        name: "difficultterrain.incrementSpeed.n",
-        hint: "difficultterrain.incrementSpeed.h",
-        scope: "client",
-        config: true,
-        default: 200,
-        type: Number
-    });
-    game.settings.register('difficultterrain', "incrementHotkey", {
-        name: "difficultterrain.incrementHotkey.n",
-        hint: "difficultterrain.incrementHotkey.h",
-        scope: "client",
-        config: true,
-        default: "+",
-        type: String
-    });
-    game.settings.register('difficultterrain', "decreaseHotkey", {
-        name: "difficultterrain.decreaseHotkey.n",
-        hint: "difficultterrain.decreaseHotkey.h",
-        scope: "client",
-        config: true,
-        default: "-",
-        type: String
-    });
-    game.settings.register('difficultterrain', "extendedRuler", {
-        name: "difficultterrain.extendedRuler.n",
-        hint: "difficultterrain.extendedRuler.h",
-        scope: "world",
-        config: true,
-        default: "DragRuler",
-        type: String
-    });
-});
+Hooks.once('init', ModuleSettings.registerSettings);
